feat: New SendableProperty wrapper to protect RW access of a property.
