+++ Sources/InfomaniakCore/Asynchronous/Sendable/SendableDictionary.swift
